+++ docs/source/conf.py
@@ -7,7 +7,7 @@
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = "PydSEAMSlib"
-release = "2.0.1"
+release = "2.1.0"
 copyright = "2023--present, d-SEAMS developers"
 author = "Ruhila S"